+++ app/src/components/dnd/ComponentList.tsx
@@ -25,8 +25,8 @@ import { SortableComponentItem } from "./@dnd/SortableComponentItem";
 import { Item } from "./@dnd/Item";
 
 // Intefaces
-import ComponentService from "../../services/component.service";
-import { Component } from "../../interfaces/SectionInfo";
+import ComponentService from "@services/component.service";
+import { Component } from "@interfaces/Course";
 
 interface Props {
   sid: string;
