+++ app/src/contexts/courseStore.tsx
@@ -1,5 +1,5 @@
 import React, { createContext, useContext, useState } from 'react';
-import { Course, Section, Lecture } from '../interfaces/Course';
+import { Course, Section, Lecture, Exercise } from '../interfaces/Course';
 
 // Define the context type
 interface CourseContextProps {
@@ -9,6 +9,8 @@ interface CourseContextProps {
   updateSections: (sections: Section[]) => void;
   lectures: Lecture[];
   updateLectures: (lectures: Lecture[]) => void;
+  Exercise: Exercise[];
+  updateExercises: (exercises: Exercise[]) => void;
 }
 
 interface CourseProviderProps {
@@ -22,13 +24,16 @@ const CourseContext = createContext<CourseContextProps>({
   sections: [],
   updateSections: () => {},
   lectures: [],
-  updateLectures: () => {}
+  updateLectures: () => {},
+  Exercise: [],
+  updateExercises: () => {},
 });
 
 export const CourseProvider: React.FC<CourseProviderProps> = ({ children }) => {
   const [course, setCourse] = useState<Course | undefined>(undefined);
   const [sections, setSections] = useState<Section[]>([]);
   const [lectures, setLectures] = useState<Lecture[]>([]);
+  const [exercises, setExercises] = useState<Exercise[]>([]);
 
   const updateCourse = (newCourse: Course) => {
     setCourse(prevCourse => ({
@@ -44,6 +49,7 @@ export const CourseProvider: React.FC<CourseProviderProps> = ({ children }) => {
 
   const updateLectures = (newLectures: Lecture[]) => {
     setLectures(prevLectures => ([...prevLectures, ...newLectures]));
+    
   };
 
   const value = {
