--- app/src/interfaces/Exercise.ts
@@ -1,9 +1,0 @@
-import {Answer} from "./Answer";
-
-export interface Exercise {
-    _id: string,
-    parentSection: string,
-    title: string,
-    question: string,
-    answers: Answer[]
-}
