--- app/src/interfaces/Lecture.ts
@@ -1,7 +1,0 @@
-export interface Lecture {
-    _id: string,
-    title: string,
-    description: string,
-    contentType: string,
-    content: string,
-}
