--- app/src/interfaces/SectionInfo.ts
@@ -1,5 +1,0 @@
-export interface Component {
-    compId: string;
-    compType: string;
-    _id: string;
-  }
